drivers/amlogic/nand/amlnf: Hard code (for now) valid discard_granularity and max_discard_sectors for the Vero 2 nand.
Without a valid non zero discard_granularity lvm will not pass down discard requests from the file system to the nand block devices, thus the nand runs out of discarded_pages and performance degrades.
Discard granularity for the nand was empirically determined to be 64 sectors. (32768 bytes) A max_discard_sectors of UINT_MAX was also found to be invalid as the ioctl fails for discard ranges larger
than about 1GB. The largest multiple of the discard_granularity that succeeds is 65535*64 = 4194240 sectors. Also added debug printk to confirm discarding is working.